+++ include/clasp/main.hpp
@@ -55,8 +55,8 @@
 #ifndef CLASP_DOCUMENTATION_SKIP_SECTION
 # define CLASP_VER_CLASP_HPP_MAIN_MAJOR     2
 # define CLASP_VER_CLASP_HPP_MAIN_MINOR     1
-# define CLASP_VER_CLASP_HPP_MAIN_REVISION  0
-# define CLASP_VER_CLASP_HPP_MAIN_EDIT      46
+# define CLASP_VER_CLASP_HPP_MAIN_REVISION  2
+# define CLASP_VER_CLASP_HPP_MAIN_EDIT      51
 #endif /* !CLASP_DOCUMENTATION_SKIP_SECTION */
 
 
@@ -146,6 +146,12 @@ namespace main
  */
 typedef int (STLSOFT_CDECL *pfnMain_t)(clasp::arguments_t const* args);
 
+/** Function pointer to a
+ * <code>clasp_main(clasp::arguments_t const& args)</code> that may be
+ * passed to clasp::main::invoke() overload.
+ */
+typedef int (STLSOFT_CDECL *pfnMainRef_t)(clasp::arguments_t const& args);
+
 
 /* /////////////////////////////////////////////////////////////////////////
  * implementation functions
@@ -157,11 +163,82 @@ namespace ximpl
 {
 
 inline
+clasp::char_t const*
+resolve_program_name_(
+    clasp::char_t const*        programName
+,   clasp::char_t const* const* argv
+)
+{
+    assert(NULL != argv);
+
+    if (NULL != programName &&
+        '\0' == programName[0])
+    {
+        programName = NULL;
+    }
+#ifdef CLASP_MAIN_DEFAULT_PROGRAM_NAME
+
+    if (NULL == programName)
+    {
+        programName = CLASP_MAIN_DEFAULT_PROGRAM_NAME;
+    }
+#endif /* CLASP_MAIN_DEFAULT_PROGRAM_NAME */
+#if !defined(PANTHEIOS_USE_WIDE_STRINGS) && \
+    defined(PANTHEIOS_INCL_PANTHEIOS_H_PANTHEIOS) && \
+    PANTHEIOS_VER >= 0x010001d6
+
+    if (NULL == programName)
+    {
+# ifndef PANTHEIOS_NO_NAMESPACE
+
+        using pantheios::pantheios_getProcessIdentity;
+# endif /* !PANTHEIOS_NO_NAMESPACE */
+
+        programName = pantheios_getProcessIdentity();
+    }
+#endif
+
+    if (NULL == programName)
+    {
+#if (STLSOFT_LEAD_VER >= 0x010a0000) && \
+    (   !defined(CLASP_USE_WIDE_STRINGS) || \
+        STLSOFT_LEAD_VER >= 0x010a0113)
+
+        programName = platformstl_ns_qual(get_executable_name_from_path)(argv[0]).ptr;
+#else
+
+        programName = argv[0];
+#endif
+    }
+
+    if (NULL == programName)
+    {
+        programName = argv[0];
+    }
+
+    return programName;
+}
+
+struct invoke_ref_adaptor_t
+{
+    pfnMainRef_t pfn;
+
+    int operator ()(clasp::arguments_t const* args) const
+    {
+        assert(NULL != args);
+        assert(NULL != pfn);
+
+        return pfn(*args);
+    }
+};
+
+template <typename F>
+inline
 int
 invoke_(
     int                                 argc
 ,   clasp::char_t const* const*         argv
-,   int (STLSOFT_CDECL*                 pfnMain)(clasp::arguments_t const* args)
+,   F                                   pfnMain
 ,   clasp::char_t const*                programName
 ,   clasp::specification_t const        specifications[]
 ,   unsigned                            flags
@@ -187,7 +264,6 @@ invoke_(
 
     assert(argc > 0);
     assert(NULL != argv);
-    assert(NULL != pfnMain);
     assert(NULL != programName);
 
     clasp::arguments_t const* clargs;
@@ -259,22 +335,93 @@ invoke_(
  * functions
  */
 
+/** \overload
+ *
+ * This overload takes the callback before the specifications array, and
+ * does not default \c programName or \c flags.
+ */
+inline
+int
+invoke(
+    int                                 argc
+,   clasp::char_t const* const*         argv
+,   pfnMain_t                           pfnMain
+,   clasp::char_t const*                programName
+,   clasp::specification_t const        specifications[]
+,   unsigned                            flags
+,   clasp::diagnostic_context_t const*  ctxt            =   NULL
+,   clasp::char_t const*                usageHelpSuffix =   NULL
+)
+{
+    assert(NULL != pfnMain);
+
+    programName = ximpl::resolve_program_name_(programName, argv);
+
+    return ximpl::invoke_(
+        argc
+    ,   argv
+    ,   pfnMain
+    ,   programName
+    ,   specifications
+    ,   flags
+    ,   ctxt
+    ,   usageHelpSuffix
+    );
+}
+
+/** \overload
+ *
+ * This overload accepts a callback that takes
+ * <code>clasp::arguments_t const&</code> rather than a pointer, with the
+ * callback before the specifications array.
+ */
+inline
+int
+invoke(
+    int                                 argc
+,   clasp::char_t const* const*         argv
+,   pfnMainRef_t                        pfnMain
+,   clasp::char_t const*                programName
+,   clasp::specification_t const        specifications[]
+,   unsigned                            flags
+,   clasp::diagnostic_context_t const*  ctxt            =   NULL
+,   clasp::char_t const*                usageHelpSuffix =   NULL
+)
+{
+    assert(NULL != pfnMain);
+
+    programName = ximpl::resolve_program_name_(programName, argv);
+
+    ximpl::invoke_ref_adaptor_t const adaptor = { pfnMain };
+
+    return ximpl::invoke_(
+        argc
+    ,   argv
+    ,   adaptor
+    ,   programName
+    ,   specifications
+    ,   flags
+    ,   ctxt
+    ,   usageHelpSuffix
+    );
+}
+
 /** Parses the command-line (specified in \c argc and \c argv) and invokes
  * caller-supplied CLASP main function (\c pfnMain) according to the given
  * arguments.
  *
  * \param argc \c argc passed to <code>main()</code>;
  * \param argv \c argv passed to <code>main()</code>;
+ * \param specifications Pointer to an specifications array that will be
+ *   passed to clasp::parseArguments();
  * \param pfnMain Caller-supplied CLASP main function that will be invoked;
  * \param programName Specifies the name of the program, which will be
  *   inferred heuristically if \c NULL or empty;
- * \param specifications Pointer to an specifications array that will be
- *   passed to clasp::parseArguments();
  * \param flags Flags that will be passed to clasp::parseArguments();
+ * \param ctxt Diagnostic context. May be \c NULL;
  * \param usageHelpSuffix Suffix such as "use --help for usage" that will be
  *   semicolon-space appended after the exception information, or \c NULL
  *   for no suffix;
- * \param ctxt Diagnostic context. May be \c NULL;
  *
  * \note If use of the Pantheios diagnostic logging API library is detected,
  *   via Pantheios C and/or C++ API main headers - pantheios/pantheios.h and
@@ -302,60 +449,15 @@ int
 invoke(
     int                                 argc
 ,   clasp::char_t const* const*         argv
-,   pfnMain_t                           pfnMain
-,   clasp::char_t const*                programName
 ,   clasp::specification_t const        specifications[]
-,   unsigned                            flags
+,   pfnMain_t                           pfnMain
+,   clasp::char_t const*                programName     =   NULL
+,   unsigned                            flags           =   0
 ,   clasp::diagnostic_context_t const*  ctxt            =   NULL
 ,   clasp::char_t const*                usageHelpSuffix =   NULL
 )
 {
-    if (NULL != programName &&
-        '\0' == programName[0])
-    {
-        programName = NULL;
-    }
-#ifdef CLASP_MAIN_DEFAULT_PROGRAM_NAME
-
-    if (NULL == programName)
-    {
-        programName = CLASP_MAIN_DEFAULT_PROGRAM_NAME;
-    }
-#endif /* CLASP_MAIN_DEFAULT_PROGRAM_NAME */
-#if !defined(PANTHEIOS_USE_WIDE_STRINGS) && \
-    defined(PANTHEIOS_INCL_PANTHEIOS_H_PANTHEIOS) && \
-    PANTHEIOS_VER >= 0x010001d6
-
-    if (NULL == programName)
-    {
-# ifndef PANTHEIOS_NO_NAMESPACE
-
-        using pantheios::pantheios_getProcessIdentity;
-# endif /* !PANTHEIOS_NO_NAMESPACE */
-
-        programName = pantheios_getProcessIdentity();
-    }
-#endif
-
-    if (NULL == programName)
-    {
-#if (STLSOFT_LEAD_VER >= 0x010a0000) && \
-    (   !defined(CLASP_USE_WIDE_STRINGS) || \
-        STLSOFT_LEAD_VER >= 0x010a0113)
-
-        programName = platformstl_ns_qual(get_executable_name_from_path)(argv[0]).ptr;
-#else
-
-        programName = argv[0];
-#endif
-    }
-
-    if (NULL == programName)
-    {
-        programName = argv[0];
-    }
-
-    return ximpl::invoke_(
+    return invoke(
         argc
     ,   argv
     ,   pfnMain
@@ -367,13 +469,20 @@ invoke(
     );
 }
 
+/** \overload
+ *
+ * This overload accepts a callback that takes
+ * <code>clasp::arguments_t const&</code> rather than a pointer, with the
+ * specifications array before the callback and defaults for \c programName
+ * and \c flags.
+ */
 inline
 int
 invoke(
     int                                 argc
 ,   clasp::char_t const* const*         argv
 ,   clasp::specification_t const        specifications[]
-,   pfnMain_t                           pfnMain
+,   pfnMainRef_t                        pfnMain
 ,   clasp::char_t const*                programName     =   NULL
 ,   unsigned                            flags           =   0
 ,   clasp::diagnostic_context_t const*  ctxt            =   NULL